requirements.txt
#torch~=1.5.1
torch~= 1.13.0
#torchvision~=0.6.1
torchvision==0.14.0
numpy~=1.15.4
Pillow~=6.1.0
pycocotools~=2.0
matplotlib~=3.0.2
tensorboardX~=2.0
